# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Golang CMS API",
    "description": "A content management system API built with Go, featuring JWT authentication with refresh tokens, password reset by email, and user profile management. Request bodies are limited to 1 MB (413 when larger).",
    "version": "1.0.0",
    "x-api-version": "v1",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/api/v1/roles": {
      "get": {
        "tags": [
          "Roles"
        ],
        "summary": "List roles",
        "description": "Needs `roles:read`.",
        "operationId": "listRoles",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Roles with their permissions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Role"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing, invalid, or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 1003,
                  "message": "Unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - the user lacks the `roles:read` permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 1004,
                  "message": "You do not have permission to perform this action"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Roles"
        ],
        "summary": "Create a role",
        "description": "Needs `roles:create`, and every permission given to the role.",
        "operationId": "createRole",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Role created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed, or an unknown permission ID (code 1002)",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ValidationErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing, invalid, or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 1003,
                  "message": "Unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - the user lacks the `roles:create` permission, or the change involves a permission the user does not hold",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 1004,
                  "message": "You do not have permission to perform this action"
                }
              }
            }
          },
          "409": {
            "description": "Role name already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 1005,
                  "message": "Role name already exists"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/roles/{id}": {
      "get": {
        "tags": [
          "Roles"
        ],
        "summary": "Get a role",
        "description": "Needs `roles:read`.",
        "operationId": "getRole",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Role ID",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The role with its permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          },
          "400": {
            "description": "Invalid id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 1002,
                  "message": "Invalid id"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing, invalid, or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 1003,
                  "message": "Unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - the user lacks the `roles:read` permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 1004,
                  "message": "You do not have permission to perform this action"
                }
              }
            }
          },
          "404": {
            "description": "Role not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 1001,
                  "message": "Role not found"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Roles"
        ],
        "summary": "Update a role",
        "description": "Needs `roles:update`, and every permission the role loses or gains. The `admin` role cannot be changed.",
        "operationId": "updateRole",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Role ID",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Role updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed, invalid id, or an unknown permission ID",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ValidationErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing, invalid, or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 1003,
                  "message": "Unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - the user lacks the `roles:update` permission, or the change involves a permission the user does not hold, or the role is `admin`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 1004,
                  "message": "You do not have permission to perform this action"
                }
              }
            }
          },
          "404": {
            "description": "Role not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 1001,
                  "message": "Role not found"
                }
              }
            }
          },
          "409": {
            "description": "Role name already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 1005,
                  "message": "Role name already exists"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Roles"
        ],
        "summary": "Delete a role",
        "description": "Needs `roles:delete`, and every permission of the role. The role is removed from its users. The `admin` role cannot be deleted.",
        "operationId": "deleteRole",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Role ID",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Role deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Delete role successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 1002,
                  "message": "Invalid id"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing, invalid, or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 1003,
                  "message": "Unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - the user lacks the `roles:delete` permission, or the change involves a permission the user does not hold, or the role is `admin`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 1004,
                  "message": "You do not have permission to perform this action"
                }
              }
            }
          },
          "404": {
            "description": "Role not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 1001,
                  "message": "Role not found"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/permissions": {
      "get": {
        "tags": [
          "Roles"
        ],
        "summary": "List permissions",
        "description": "Needs `permissions:read`.",
        "operationId": "listPermissions",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "All permissions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Permission"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing, invalid, or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 1003,
                  "message": "Unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - the user lacks the `permissions:read` permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 1004,
                  "message": "You do not have permission to perform this action"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{id}/roles": {
      "put": {
        "tags": [
          "Roles"
        ],
        "summary": "Replace the roles of a user",
        "description": "Needs `users:assign-roles`, and every permission of the roles the user gains or loses. The last admin cannot lose the `admin` role.",
        "operationId": "setUserRoles",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "User ID",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetUserRolesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The user's roles after the change",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Role"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed, invalid id, or an unknown role ID",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ValidationErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing, invalid, or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 1003,
                  "message": "Unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - the user lacks the `users:assign-roles` permission, or the change involves a permission the user does not hold, or it would remove the last admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 1004,
                  "message": "You do not have permission to perform this action"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": 1001,
                  "message": "User not found"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "Application error code (see pkg/apperror/codes.go)",
            "example": 1003
          },
          "message": {
            "type": "string",
            "example": "Unauthorized"
          }
        }
      },
      "ValidationErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 4001
          },
          "message": {
            "type": "string",
            "example": "Validation failed"
          },
          "fields": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string",
                  "example": "email"
                },
                "message": {
                  "type": "string",
                  "example": "email is required"
                }
              }
            }
          }
        }
      },
...
      "Permission": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "roles:read"
          },
          "description": {
            "type": "string",
            "example": "View roles"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Role": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 2
          },
          "name": {
            "type": "string",
            "example": "editor"
          },
          "description": {
            "type": "string",
            "example": "Can edit settings"
          },
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Permission"
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "RoleRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "example": "editor"
          },
          "description": {
            "type": "string",
            "maxLength": 255,
            "example": "Can edit settings"
          },
          "permission_ids": {
            "type": "array",
            "items": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Replaces the role's permissions; omit or leave empty for none",
            "example": [
              7,
              8
            ]
          }
        }
      },
      "SetUserRolesRequest": {
        "type": "object",
        "required": [
          "role_ids"
        ],
        "properties": {
          "role_ids": {
            "type": "array",
            "items": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Replaces all roles of the user; an empty list removes every role",
            "example": [
              2
            ]
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token in the Authorization header using the Bearer scheme"
      }
    }
  }
}